refactor(voyager): tear down web lib/ into api/services/constants (R-220, R-240)
Moves business-logic modules to services/, the fetch transport to api/request.ts, shared constants to constants.ts, and folds context/+providers/ into state/; rewrites all @/ alias imports.
