+++ apps/client/web/src/app/(auth)/login/page.tsx
@@ -10,10 +10,10 @@ import { type FormEvent, useState } from 'react';
 import Link from 'next/link';
 import { useRouter, useSearchParams } from 'next/navigation';
 
+import { ApiError, get } from '@/api/request';
 import { Toast } from '@/components/Toast/Toast';
-import { useAuth } from '@/context/AuthContext';
-import { ApiError, get } from '@/lib/api';
-import { APP_NAME } from '@/lib/constants';
+import { APP_NAME } from '@/constants';
+import { useAuth } from '@/state/AuthContext';
 
 import styles from '../auth.module.scss';
