+++ apps/client/web/src/app/(protected)/trips/new/page.tsx
@@ -9,7 +9,7 @@ import { useEffect, useRef, useState } from 'react';
 import Link from 'next/link';
 import { useRouter, useSearchParams } from 'next/navigation';
 
-import { post } from '@/lib/api';
+import { post } from '@/api/request';
 
 import styles from './newTrip.module.scss';
