+++ apps/client/web/src/app/layout.tsx
@@ -10,9 +10,9 @@ import 'mapbox-gl/dist/mapbox-gl.css';
 import { ErrorBoundary } from '@/components/ErrorBoundary/ErrorBoundary';
 import { Footer } from '@/components/Footer/Footer';
 import { Header } from '@/components/Header/Header';
-import { AuthProvider } from '@/context/AuthContext';
-import { APP_NAME } from '@/lib/constants';
-import { QueryProvider } from '@/providers/QueryProvider';
+import { APP_NAME } from '@/constants';
+import { AuthProvider } from '@/state/AuthContext';
+import { QueryProvider } from '@/state/QueryProvider';
 
 import '../styles/animations.scss';
 import './globals.scss';
