+++ apps/client/web/src/components/AuthGuard/AuthGuard.tsx
@@ -9,7 +9,7 @@ import { useEffect } from 'react';
 
 import { usePathname, useRouter, useSearchParams } from 'next/navigation';
 
-import { useAuth } from '@/context/AuthContext';
+import { useAuth } from '@/state/AuthContext';
 
 export function AuthGuard({ children }: { children: React.ReactNode }) {
   const { user, isLoading } = useAuth();
