+++ apps/client/web/src/components/ChatBox/VirtualizedChat.tsx
@@ -10,7 +10,7 @@ import { useCallback, useEffect, useMemo, useRef } from 'react';
 import type { ChatMessage, ChatNode, TripPlanCard } from '@repo/types';
 import { useVirtualizer } from '@tanstack/react-virtual';
 
-import { APP_NAME } from '@/lib/constants';
+import { APP_NAME } from '@/constants';
 
 import { NodeRenderer } from './NodeRenderer';
 import styles from './VirtualizedChat.module.scss';
