+++ apps/client/web/src/components/ChatBox/useSSEChat.ts
@@ -11,7 +11,7 @@ import { useCallback, useEffect, useRef, useState } from 'react';
 import type { ChatNode, SSEEvent } from '@repo/types';
 import { useQueryClient } from '@tanstack/react-query';
 
-import { API_BASE } from '@/lib/api';
+import { API_BASE } from '@/api/request';
 
 interface UseSSEChatOptions {
   tripId: string;
