+++ apps/client/web/src/components/ChatBox/widgets/ExperienceCard.tsx
@@ -4,8 +4,8 @@
  * selection groups, rendering as a pressable button for keyboard and screen
  * reader access.
  */
-import { API_BASE } from '@/lib/api';
-import { formatCurrency } from '@/lib/format';
+import { API_BASE } from '@/api/request';
+import { formatCurrency } from '@/services/format';
 
 import styles from './ExperienceCard.module.scss';
 import { MapPreviewCard } from './MapPreviewCard';
