+++ apps/client/web/src/components/ChatBox/widgets/InlineBudgetBar.tsx
@@ -3,7 +3,7 @@
  * over-budget state. Exists to give an at-a-glance cost signal inline within
  * chat cards without the full budget widget.
  */
-import { formatCurrency } from '@/lib/format';
+import { formatCurrency } from '@/services/format';
 
 import styles from './InlineBudgetBar.module.scss';
