+++ apps/client/web/src/components/Header/Header.tsx
@@ -12,10 +12,10 @@ import { usePathname } from 'next/navigation';
 import * as Dialog from '@radix-ui/react-dialog';
 import { useQuery } from '@tanstack/react-query';
 
-import { useAuth } from '@/context/AuthContext';
-import { get } from '@/lib/api';
-import { APP_NAME, GITHUB_REPO_URL } from '@/lib/constants';
-import { type UserPreferences } from '@/lib/preferenceOptions';
+import { get } from '@/api/request';
+import { APP_NAME, GITHUB_REPO_URL } from '@/constants';
+import { type UserPreferences } from '@/services/preferenceOptions';
+import { useAuth } from '@/state/AuthContext';
 
 import styles from './Header.module.scss';
