+++ apps/client/web/src/components/PreferencesWizard/steps/AccommodationStep.tsx
@@ -3,7 +3,7 @@
  * ACCOMMODATION_OPTIONS as a single-select so the questionnaire captures lodging
  * style in one place.
  */
-import { ACCOMMODATION_OPTIONS } from '@/lib/preferenceOptions';
+import { ACCOMMODATION_OPTIONS } from '@/services/preferenceOptions';
 
 import styles from '../PreferencesWizard.module.scss';
