+++ apps/client/web/src/components/PreferencesWizard/steps/ActivitiesStep.tsx
@@ -3,7 +3,7 @@
  * multi-select so the questionnaire can capture several preferred activity types
  * at once.
  */
-import { ACTIVITY_OPTIONS } from '@/lib/preferenceOptions';
+import { ACTIVITY_OPTIONS } from '@/services/preferenceOptions';
 
 import styles from '../PreferencesWizard.module.scss';
