+++ apps/client/web/src/components/PreferencesWizard/steps/DiningStep.tsx
@@ -3,7 +3,10 @@
  * single-select dining style so the questionnaire captures both constraints and
  * taste in one combined step.
  */
-import { DIETARY_OPTIONS, DINING_STYLE_OPTIONS } from '@/lib/preferenceOptions';
+import {
+  DIETARY_OPTIONS,
+  DINING_STYLE_OPTIONS,
+} from '@/services/preferenceOptions';
 
 import styles from '../PreferencesWizard.module.scss';
