+++ apps/client/web/src/state/AuthContext.tsx
@@ -9,7 +9,7 @@ import { createContext, useCallback, useContext, useMemo } from 'react';
 
 import { useQuery, useQueryClient } from '@tanstack/react-query';
 
-import { ApiError, get, post } from '@/lib/api';
+import { ApiError, get, post } from '@/api/request';
 
 export interface User {
   id: string;
